Add base image for Debian9 qt4 version
Provides a Docker file to generate a base image based
on Debian9. It includes all requirements for mxcube2
with qt4 (current master branch).

The image is primarily intended to mount your local copy of the
mxcube repository for development purposes. However, the same
image can also be used as base image for the demo version.

We also include a script to start the docker container.