Cleaned up imports, and bug fixes in latest merge
Regrettably the import cleanup (or a small part of it) was required
for the bug fix, hence the mixing of the two.

The import changes have moved to importing from HardwareRepository. etc.
and to gui. etc. in all cases, avoiding direct imports of e.g.
sample_centring, abstract.xyz, individual HOs etc.

The reason is that importing the same file under two different paths
creates two different copies in memory, so that state_x == state_y
or setting class attributes gives unpredictable results when the
class/instance is not from the same package in different cases.